// src/shared/swagger/swagger.interface.ts
export interface SwaggerOptions {
  username: string
  password: string
}

export enum HttpStatusCode {
  OK = 200,
  CREATED = 201,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  CONFLICT = 409,
  INTERNAL_SERVER_ERROR = 500,
}

export interface ResponseMessage {
  statusCode: number
  message: string
  description?: string
}